osd/osd_types: remove cat_sum from object_stat_collection_t
This type is now just a useless wrapper around object_stat_sum_t, but that
is not easy to drop.

Note that now if we load and reencode a legacy object we'll drop those
fields.  We would mark the incompat flag in the object corpus accordingly,
but nothing in there populates the cat fields.

Signed-off-by: Sage Weil <sage@redhat.com>