Revert "osd: detect (some) misordered ondisk tmaps"
This reverts commit f91e98b51dc93635c079f3958b7bbdbea872589d.

The check for 'nextkey < last_disk_key' makes not much sense since
last_disk_key is an empty string and not set before. Comparing a
decoded string to be less than an empty string will be never true.

Since this if() isn't part of a loop last_disk_key is only set
once and there is no other consumer: revert this dead code.

Conflicts:
	src/osd/ReplicatedPG.cc