crush/CrushWrapper: fix detach_bucket
In commit 7ebd208546d94b9499919427262ab9fed30ea94d we changed the call that
changed the weight of all instances of item to one that explicitly
changes it in the parent bucket, but parent_id may not be valid at the
call site.  Move this into the conditional block to fix.

Fixes: #10095
Signed-off-by: Sage Weil <sage@redhat.com>