Various bug fixes while running tests on this branch in the octo lab.
Made YamlConfig work more closely like a dictionary by extending
collections.MutableMapping.  It still needs some work to be able to be
used interchangeably with a dict - most noteably with yaml.safe_dump.

Decided to just use a dict instead of JobConfig for args['<config>'].
Too many other codepaths are assuming that config is a dict.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>