crush: add allowed_bucket_types tunable
This tunable is a bitmask indicating which bucket types are allowed.
For now, the only purpose is to affect get_default_bucket_type(), which
will try to pick a type that is supported, with a preference for straw2 or
straw buckets.  In the future, we likely want something a bit more
sophisticated that reflects whether the bucket is expected to be fixed-size
or not.

We could have added a default_bucket_type field to accomplish this, but I
think that would prove to be too limiting in the future, and this
accomplishes the same thing.

Note that if the admin selects the 'hammer' tunables, that means clients
and servers will support straw2 and that will be the preferred choice. The
default is still bobtail tunables, which means new clusters created on
hammer will not use straw2 just yet.

Signed-off-by: Sage Weil <sage@redhat.com>