osd/ReplicatedPG: introduce new methods for tracking SnapSet updates
Explicitly indicate when the SnapSet is modified, when we must force a
write (despite maybe not being updated), and a helper to tell if it should
be written.  Also provide helpers to setting SnapSet fields; often these
are no-ops and we want to avoid needless attr updates in that case.

Not used yet.

Signed-off-by: Sage Weil <sage@redhat.com>