osd/ReplicatedPG: apply snapset helpers in finish_ctx
Make sure we write the snapset for a new object or when expclitly
requested.  Rejigger the attr encoding a bit since we don't know upfront
whether we will want to write it or not (if we do decide we need it,
encode it at the same time).  Note that !bss.empty() ==
should_write_snapset() is effectively the invariant we are after.

This is originally based on Xinze Chi <xmdxcxz@gmail.com>'s patch.

Signed-off-by: Sage Weil <sage@redhat.com>