Don't assume y on distro kernel reinstall...
Because if its the running kernel yum will skip the current
version and install install an older package? WTF? I would not
expect such behavior from a reinstall but it is what it is...

Also skipping current running kernel output was in stderr not the
stdout so I made changes to reflect that.

Signed-off-by: Sandon Van Ness <sandon@redhat.com>