Revert "ceph.spec.: add epoch"
If ICE ships 0.80.8, then it will be newer than what RHCEPH ships
(0.80.7), and users won't be able to seamlessly upgrade via Yum.

We have three options:
A) Revert the "Epoch: 1" change on the Firefly branch.
B) Revert the "Epoch: 1" change in the ICE packages.
C) Bump the Epoch to "2" in Red Hat's packages.

This commit does Option A.

Option B may or may not be feasible - it would require a "downstream"
change in ICE, and we haven't done that sort of thing before.

Due to the RHEL release schedule, Option C is not available to us at
this point.

This reverts commit b890c1e4706d7cfef7ed24c9df65b439b4f7ff1d.