tasks/cephfs: let get_config use non-mon services
Previously was always using the default values of things
so querying mon instead of the appropriate service
worked fine.  However, for things we might want to
update on a per-test basis we need to go ask the
correct service what the setting really is.

Needed for osd_mon_report_interval_max in the ENOSPC
testing.

Signed-off-by: John Spray <john.spray@redhat.com>