Use teuthology.misc.get_distro and teuthology.misc.get_distro_ version in the internal.lock_machines task.
When the internal task wasn't using get_distro to retrieve os_type it
could set os_type to None, causing the lock to fail. This happened when
there was not an os_type defined in the config or a --os-type flag given.
Now, os_type will be properly defaulted to ubuntu when trying to lock
bare metal machines.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>