Allow os_version and os_type to be None when locking bare metal nodes.
This fixes a bug that was defaulting os_version to 12.04 and os_type to
ubuntu when locking bare metal machines.  We don't want to default the
os_version or os_type when locking bare metal so that the locking code
can pick whichever node is available.

VPS nodes will get proper defaults in provision.create_if_vm if they
aren't explicitly defined by either the config yaml or the cli flags.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>