osd: requeue PG when we skip handling a peering event
If we don't handle the event, we need to put the PG back into the peering
queue or else the event won't get processed until the next event is
queued, at which point we'll be processing events with a delay.

The queue_null is not necessary (and is a waste of effort) because the
event is still in pg->peering_queue and the PG is queued.

This is a firefly-specific patch; a (very) similar one will appear in master
in 492ccc900c3358f36b6b14a207beec071eb06707.

Backport: giant, firefly
Signed-off-by: Sage Weil <sage@redhat.com>