osd: enable filestore_extsize by default
Note that this will only get used if the kernel is new enough; if it is
older than 3.5 the option will get disabled and extsize will not be used
even if the option is set to true.

This partially reverts 30d16ac015ab392fb2b32ec4d1e6e3f4d9a145cb.

Fixes: #9956
Signed-off-by: Sage Weil <sage@redhat.com>