Add crush_ln to calculate nature log efficently
Crush_ln is a function that could calculate ln(x+1) * 4096, which is exactly maps [0,0xffff]
(corresponding to real numbers [1/0x10000, 1] to [0, 0xffff] (corresponding to real numbers
[-11.090355,0]).

It is architecture-independent, and runs in 22 to 52 cycles on a Core Duo (faster for longer inputs, slower for short inputs;
call overhead is not included).

It matches the crush_ln_table outputs in most cases, with differences of +/-1 in 4.35% of cases.

Signed-off-by: Xiaoxi Chen <xiaoxi.chen@intel.com>
Signed-off-by: Anderson, Cristina S <cristina.s.anderson@intel.com>