auth/cephx: optimize signature check
The encode_encrypt() helper will generate a bufferlist with a simple
structure.  Profiles seem to indicate this is taking a large amount of time
in the message receive path.

Avoid the encode overhead since we have a tiny buffer with a fixed and
known structure.

Reported-by: Andreas Bluemle <andreas.bluemle@itxperts.de>
Signed-off-by: Sage Weil <sage@redhat.com>
(cherry picked from commit 8d16d4ce14a82356007c14fb7535170933eb7812)