librados: add FADVISE_NOREUSE
We left this off because it seemed the same as DONTNEED, but there is a
subtle distinction: DONTNEED means nobody will need it (and we probably
discard our cache), while NOREUSE means this client won't need it again
(and we should try to avoid polluting the cache from this IO only).  At
least, that's the way we'r defining it.  posix_fadvise says:

       POSIX_FADV_NOREUSE
              The specified data will be accessed only once.
       POSIX_FADV_DONTNEED
              The specified data will not be accessed in the near future.

which is similar.  I think our definitions make a bit more sense for the
multi-client environment.

Signed-off-by: Sage Weil <sage@redhat.com>