ReplicatedPG::hit_set_persist: update ssc->snapset as well
This is a hack.  The correct answer is to adapt this method and
finish_ctx to allow this method to use finish_ctx.  This is complicated
by the presence of the hit set trims in the same repop, so for now, we
kick the can down the road a bit.

Fixes: 10616
Signed-off-by: Samuel Just <sjust@redhat.com>