librbd: trim header update not using AIO
The original trim header update code was using blocking IO to
update the header.  After migrating to an asynchronous trim
which performs all work in librados callbacks, it exposed a
potential deadlock in the librados_test_stub when attempting
to do blocking IO within a librados callback.  This commit
changes the header update to use AIO.

Fixes: #10637
Signed-off-by: Jason Dillaman <dillaman@redhat.com>