librados: handle failed watch registration better
If watch() failed, the LingerOp would be deleted, but a pointer to it
would live on in the handle passed by the caller. An unwatch() called
on that handle would cause a segfault.

Set handle only after the watch osd op has returned, and set it to 0
on failure. Make unwatch() return 0 early if passed a handle of 0, to
get back the earlier behavior of not segfaulting if a non-registered
cookie is passed to unwatch(). This aspect of the interface changed
when the new watch2/notify2 functions were introduced.

Fixes #10644
Signed-off-by: Josh Durgin <jdurgin@redhat.com>