common: filtering in `perf dump`
So that we can get out a particular subsystem
or particular counter without dumping
everything.  Should make tests that watch perf
counters much less spammy!

`logger` and `counter` params are used with
an exact comparison here but the interface
should be amenable to extending to e.g. globbing
if we wanted to in the future.

Also tidy up some iterator while()s into for()s.

Signed-off-by: John Spray <john.spray@redhat.com>