osd: preserve extra_reqids in PGLog
The log scraping is unfortunately O(n) in the size of the log.  We can
also follow objects and then the prior_version pointers, but we don't have
a version->entry index at the moment.

Only index by reqid if reqid_is_index.  Always index extra_reqids.
Notably, the PROMOTE op does not return true for reqid_is_indexed, but this
is where the promotion extra_reqids will appear.

Signed-off-by: Sage Weil <sage@redhat.com>