Moved task.tests into it's own module and created tests.test_run
This sets up the basic structure for teuthology integration testing.
Any tests put in teuthology/task/tests will be autodiscovered and run
when 'tests' is in the task list of a job's config.

Each test will be given ctx and config and it's output will be logged
using our logger.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>