osdc/Objecter: do watch/notify op completions synchronously
Normally we do Op completions under completion_lock but not the
session lock.  This is problematic because the register op modifies
LingerOp state that is no longer protected by s->lock against
races from, say, _linger_cancel, or a new thread sending a fresh
register.  See #10765 for a crash that is possibly caused by this.

Fix this by making a synchronous commit hook and using that while
still holding s->lock.

Signed-off-by: Sage Weil <sage@redhat.com>