ceph.spec.in: move ceph-rbdnamer to librbd1 RPM
Prior to this commit, the RPMs and DEBs were slightly mis-aligned
regarding ceph-rbdnamer. The /usr/bin/ceph-rbdnamer utilitiy shipped in
the main "ceph" RPM, whereas in the Debian packages, the ceph-rbdnamer
utility shipped in the "librbd1" subpackage instead.

Given that:
1) the 50-rbd.rules file already ships in the librbd1 RPM, and
2) 50-rbd.rules requires /usr/bin/ceph-rbdnamer in order to function,

it makes sense to follow the Debian packages here and put
/usr/bin/ceph-rbdnamer and the ceph-rbdnamer(8) man page into the
librbd1 RPM as well.

http://tracker.ceph.com/issues/10864 Refs: #10864

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>