osd,mon: explicitly specify OSD features in MOSDBoot
We are using the connection features to populate the features field in the
OSDMap, but this is the *intersection* of mon and osd features, not the
osd features.  Fix this by explicitly specifying the features in
MOSDBoot.

Fixes: #10911
Backport: giant, firefly
Signed-off-by: Sage Weil <sage@redhat.com>
(cherry picked from commit bee91548b07c6a28314dddc50fef8b6a2677e774)

Conflicts:
	src/osd/OSD.cc
Minor difference in the MOSDBoot constructor