Client: support using dentry invalidation callbacks on older kernels
This brings back a few small code chunks that were removed in
0827bb79ea5127e6763f6e904dfa1a3266046ffb. We check the kernel version,
and if it is less than 3.18 we use these dentry invalidation callbacks
instead of the remount callback. This should resolve a number of
issues with racing against remount, including #10916, and lets older
unprivileged users on older kernels run even if they can't apply
options on mount (#10542).

Signed-off-by: Greg Farnum <gfarnum@redhat.com>